+++ FluentSecurity.Specification/Internals/HttpContextRequestDescriptionSpec.cs
@@ -1,9 +1,10 @@
 using System.Web;
 using System.Web.Routing;
+using FluentSecurity.Internals;
 using FluentSecurity.Specification.Helpers;
 using NUnit.Framework;
 
-namespace FluentSecurity.Specification
+namespace FluentSecurity.Specification.Internals
 {
 	[TestFixture]
 	[Category("HttpContextRequestDescriptionSpec")]
@@ -53,7 +54,7 @@ namespace FluentSecurity.Specification
 			var requestDescription = new HttpContextRequestDescription();
 
 			// Assert
-			Assert.That(requestDescription.AreName, Is.EqualTo("AreaName"));
+			Assert.That(requestDescription.AreaName, Is.EqualTo("AreaName"));
 			Assert.That(requestDescription.ControllerName, Is.EqualTo("ControllerName"));
 			Assert.That(requestDescription.ActionName, Is.EqualTo("ActionName"));
 		}
@@ -69,7 +70,7 @@ namespace FluentSecurity.Specification
 			var requestDescription = new HttpContextRequestDescription();
 
 			// Assert
-			Assert.That(requestDescription.AreName, Is.EqualTo(""));
+			Assert.That(requestDescription.AreaName, Is.EqualTo(""));
 			Assert.That(requestDescription.ControllerName, Is.EqualTo("ControllerName"));
 			Assert.That(requestDescription.ActionName, Is.EqualTo("ActionName"));
 		}