+++ FluentSecurity.Specification/ServiceLocation/ServiceLocatorSpec.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Configuration;
 using System.Linq;
+using FluentSecurity.Diagnostics;
+using FluentSecurity.Internals;
 using FluentSecurity.Policy.ViolationHandlers;
 using FluentSecurity.ServiceLocation;
 using FluentSecurity.Specification.Helpers;