+++ FluentSecurity/Configuration/AdvancedConfiguration.cs
@@ -1,5 +1,6 @@
 using System;
 using FluentSecurity.Caching;
+using FluentSecurity.Internals;
 using FluentSecurity.Policy.ViolationHandlers.Conventions;
 
 namespace FluentSecurity.Configuration
@@ -31,10 +32,10 @@ namespace FluentSecurity.Configuration
 			SecurityContextModifyer = modifyer;
 		}
 
-		public void Violations(Action<ViolationConfigurationExpression> violationConfigurationExpression)
+		public void Violations(Action<ViolationConfiguration> violationConfiguration)
 		{
-			if (violationConfigurationExpression == null) throw new ArgumentNullException("violationConfigurationExpression");
-			violationConfigurationExpression.Invoke(new ViolationConfigurationExpression(Conventions));
+			if (violationConfiguration == null) throw new ArgumentNullException("violationConfiguration");
+			violationConfiguration.Invoke(new ViolationConfiguration(Conventions));
 		}
 	}
 }